♻️ 调整通知相关代码
