+++ src-tauri/src/commands/mod.rs
@@ -18,7 +18,7 @@ pub use clipboard::copy_image;
 pub use front_logger::front_logger;
 pub use is_debug::is_debug;
 pub use local_storage::{get_item, set_item};
-pub use monitor_info::{get_monitor_info, message_beep};
+pub use monitor_info::message_beep;
 pub use preview_page::{back_preview, read_detail};
 pub use quit::quit;
 pub use request_client::send_request;
