+++ src-tauri/src/commands/preview_page.rs
@@ -4,8 +4,7 @@ use std::thread::sleep;
 use std::time::Duration;
 use tauri::http::{Request, Response};
 use tauri::{
-    command, AppHandle, Emitter, Manager, WebviewUrl, WebviewWindow, WebviewWindowBuilder, Window,
-    WindowBuilder, WindowEvent,
+    command, AppHandle, Emitter, Manager, WebviewUrl, WebviewWindow, WebviewWindowBuilder, WindowEvent,
 };
 use tracing::{info, instrument};
